+++ src/js/app.js
@@ -1,10 +1,9 @@
 import log from './log';
 import(/* webpackChunkName: "deferred" */ './deferred').then((module) => module.default());
 
-// Uncomment this after `html-bundler-webpack-plugin` is fixed
-// import(/* webpackChunkName: "prefetched", webpackPrefetch: true */ './prefetched').then(({ default: run }) => {
-//   run();
-// });
+import(/* webpackChunkName: "prefetched", webpackPrefetch: true */ './prefetched').then(({ default: run }) => {
+  run();
+});
 
 import(/* webpackChunkName: "prefetched-delayed", webpackPrefetch: true */ './prefetched-delayed').then(
   ({ default: run }) => {