+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 11 April 2024
+
+- Add `serve:*` npm tasks
+- Add chunk file names to `html-bundler-webpack-plugin` config
+- Update readme to not mention css extraction plugin anymore, as it is not used
+
 ## 10 April 2024
 
 - Added sourceMaps (dev, prod)
@@ -5,7 +11,7 @@
 - Changed `console.log` copy
 - Small cleanup in webpack config
 
-* `html-bundler-webpack-plugin`
+- `html-bundler-webpack-plugin`
   - Always inline entry points (JS and CSS)
   - No longer emits inlined files
   - Improved readme to explain current behavior and how to customize it